requirements.txt
python-dotenv==0.21.0
aiogram==2.22.2
requests==2.25.1
folium==0.13.0
google==3.0.0